nexusvault/requirements.txt
langgraph
langchain
langchain-community
playwright
gradio
python-dotenv
ipython
playwright 
beautifulsoup4
lxml 
html5lib
langchain-openai
langchain-experimental
wikipedia